+++ tests/test_simulate_analyze.py
@@ -8,19 +8,27 @@ from minis_validation import simulation, analysis
 TEST_DATA_DIR = Path(__file__).parent / 'data'
 
 
-@pytest.fixture(scope='session')
-def simulation_output():
-    output = TEST_DATA_DIR / 'output'
-    shutil.rmtree(output, ignore_errors=True)
-    simulation.run(TEST_DATA_DIR / 'BlueConfig',
-                   TEST_DATA_DIR / 'frequencies.csv',
-                   TEST_DATA_DIR / 'job-configs',
-                   output,
-                   mpi=False,
-                   num_cells=2,
-                   target_file=TEST_DATA_DIR / 'user.target')
+def test_simulate_analyze(tmp_path):
+    output = tmp_path / "output"
+    log_dir = tmp_path / "log"
+
+    simulation.run(
+        TEST_DATA_DIR / 'simulation_config.json',
+        TEST_DATA_DIR / 'frequencies.csv',
+        TEST_DATA_DIR / 'job-configs',
+        output,
+        num_cells=2,
+        log_dir=str(log_dir),
+        timeout_s=1200,
+        slurm_args={},
+    )
+    _print_job_logs(log_dir)
+
     analysis.analyze_jobs(TEST_DATA_DIR / 'job-configs', output)
-    return output
+
+    assert (output / 'job_results.csv').is_file()
+    _assert_job_output(output / 'PC_Exc')
+    _assert_job_output(output / 'PC_Inh')
 
 
 def _assert_job_output(job_output_dir):
@@ -32,7 +40,16 @@ def _assert_job_output(job_output_dir):
     assert (job_output_dir / 'analysis' / 'frequencies.png').is_file()
 
 
-def test_simulate_analyze(simulation_output):
-    assert (simulation_output / 'job_results.csv').is_file()
-    _assert_job_output(simulation_output / 'PC_Exc')
-    _assert_job_output(simulation_output / 'PC_Inh')
+def _print_job_logs(log_dir):
+    print("=== Standard output of jobs ==")
+    standard_paths = log_dir.glob('**/*.out')
+    for standard_path in standard_paths:
+        with open(standard_path, 'r') as f:
+            print(f.read(), end="")
+            print("=" * 30)
+    print("==== Error output of jobs ====")
+    error_paths = log_dir.glob('**/*.err')
+    for error_path in error_paths:
+        with open(error_path, 'r') as f:
+            print(f.read(), end="")
+            print("=" * 30)